requirements.txt
asttokens==3.0.0
certifi==2025.4.26
charset-normalizer==3.4.1
contourpy==1.3.0
crypto==1.4.1
cycler==0.12.1
decorator==5.2.1
exceptiongroup==1.2.2
executing==2.2.0
fonttools==4.57.0
idna==3.10
importlib-resources==6.5.2
ipython==8.18.1
jedi==0.19.2
kiwisolver==1.4.7
matplotlib==3.9.4
matplotlib-inline==0.1.7
mpmath==1.3.0
naked==0.1.32
numpy==2.0.2
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
prompt-toolkit==3.0.51
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pyyaml==6.0.2
requests==2.32.3
scipy==1.13.1
shellescape==3.8.1
six==1.17.0
stack-data==0.6.3
sympy==1.13.3
traitlets==5.14.3
typing-extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13
zipp==3.21.0